requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.8.0
Pillow==11.2.1
qrcode==8.2
